+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Segmentation module: change output file naming (do not create a segmentation_masks/ sub-folder, add _vSM suffix instead of _mask suffix).
 * Segmentation module: if input image contains multiple fields of view (F axis), raise an error instead of saving one image per field of view.
 * Cell tracking module: change output file naming (do not create a cell_tracking/ sub-folder, add _vTG suffix instead of _mask and _graph suffixes).
+* Graph filtering module: change output file naming (do not create a graph_filtering/ sub-folder, add _vGF suffix instead of _mask and _graph suffixes).
 
 ### Removed